Add person data to assignable notifications
- Move the get_person_dict function to main module so it can be used
  anyweher.
- Update refernces to that funciton.
- Add person data to request notification dictionaries.
